[ARM] tegra: clock: Handle dvfs missing regulator better
Convert the clk_is_dvfs check to only check for a dvfs structure,
and not for the dvfs regulator, so that clk_is_dvfs always returns
the same value even when CONFIG_REGULATOR=n.  Check for a missing
dvfs table in dvfs_set_rate to prevent a panic if the table is
incorrectly configured.  Move the regulator_get call in dvfs_init
to after the table configuration, so that a NULL regulator, which
is valid if CONFIG_REGULATOR=n, still initializes the table and
ensures dvfs_set_rate returns 0 for valid frequencies.

Signed-off-by: Colin Cross <ccross@android.com>