[ARM] tegra: suspend: add support for LP0 and LP1 suspend modes
in both LP0 and LP1, SDRAM is placed into self-refresh. in order to safely
perform this transition, the final shutdown procedure responsible for

  * turning off the MMU and L1 data cache
  * putting memory into self-refresh
  * setting the DDR pads to the lowest power state
  * and turning off PLLs

is copied into IRAM (at the address TEGRA_IRAM_BASE + SZ_4K) at the
start of the suspend process.

in LP1 mode (like LP2), the CPU is reset and executes the code specified
at the EVP reset vector. since SDRAM is in self-refresh, this code must
also be located in IRAM, and it must re-enable DRAM before restoring the
full context. in this implementation, it enables the CPU on PLLP, enables
PLLC and PLLM, restores the SCLK burst policy, and jumps to the LP2 reset
vector to restore the rest of the system (MMU, PLLX, coresite, etc.). the
LP2 reset vector is expected to be found in PMC_SCRATCH1, and is
initialized during system-bootup

in LP0 mode, the core voltage domain is also shutoff. as a result, all
of the volatile state in the core voltage domain (e.g., pinmux registers,
clock registers, etc.) must be saved to memory so that it can be restored
after the system resumes. a limited set of wakeups are available from LP0,
and the correct levels for the wakeups must be programmed into the PMC
wakepad configuration register prior to system shutdown. on resume, the
system resets into the boot ROM, and the boot ROM restores SDRAM and other
system state using values saved during kernel initialization in the PMC
scratch registers.

resuming from LP0 requires the boot ROM to supply a signed recovery codeblob
to the kernel; the kernel expects that the length and address of this blob
is supplied with the lp0_vec= command line argument; if not present, suspend-
to-LP0 will be disabled

for simplicity, the outer cache is shutdown for both LP0 and LP1; it
is possible to optimize the LP1 routine to bypass outer cache shutdown
and restart

v2 fixes from Vik Kasivajhula <tkasivajhula@nvidia.com>:
* restore PLLC during LP1 resume
* fix typo which set the CPU clock burst policy to PLLM, rather than PLLP

v3 fixes from Vik Kasivajhula <tkasivajhula@nvidia.com>:
* shut off the MMU during LP2, to avoid races between the flow controller
  and the instruction prefetcher

v3 fixes from Scott Williams <scwilliams@nvidia.com>:
* separate CPU-specific code from SoC-specific code

v4 fixes from Alex Frid (afrid@nvidia.com):
* refactor PMC modes to avoid dependency on polarity of external pull-ups/
  pull-downs for the power request lines
* save/restore CPU power-off timer

v4 fixes from Bharat Nihalani (bnihalani@nvidia.com):
* program deep power down pad sampling

Change-Id: I17fe667ee3fb5cf68084f51dad555ff9d5c84024
Signed-off-by: Gary King <gking@nvidia.com>