video: tegra: add utility function to compute framebuffer stride
framebuffers will generally not be exactly width * bpp / 8 bytes wide;
on tegra, linearly-addressed framebuffers will generally be rounded
up so that the stride is a multiple of 16B (so that they are compatible
with rendering from the hardware engines), and tiled framebuffers
will be a multiple of the tile width (64B).

add a utility function to tegra_dc to compute the correct stride
given a width, bpp and pixel layout, and use this in set_par.

Change-Id: I1dfcc760172ed2bf374c63d505fc6a650cbf38b9
Signed-off-by: Gary King <gking@nvidia.com>