[ARM] add work-around for 742230 Cortex-A9 erratum
Under rare circumstances, a DMB instruction between 2 writes
may not ensure the correct visibility ordering of the 2 writes
to other observers. This is most likely to reproduce by code
taking and releasing a semaphore, with the result that an
external agent may observe the semaphore release before the
payload data is visible.

The work-around disables DMB_lite operation by setting bit 4
of the undocumented diagnostic register. This causes DMB
operations to behave like DSB.

Change-Id: I22e6b2e2326d491113169d7cce52c0a33ab4f0c6
Signed-off-by: Gary King <gking@nvidia.com>