[ARM] tegra: dma: Get rid of shadowed registers
The use of shadow registers caused memory corruption at physical address
0 because the enable bit was not shadowed, and assuming it needed to be set
would enable an unconfigured dma block.  Most of the register accesses don't
need to know the previous state of the registers, and the few places that do
need to modify only a few bits in the registers are the same ones that were
sometimes incorrectly setting the enable bit.  This patch convert
tegra_dma_update_hardware to set the entire register, and the other users
to read-modify-write, and drops the shadow registers completely.

Change-Id: I846cf908c17e270a30833862712a6cff0265f197
Signed-off-by: Colin Cross <ccross@android.com>