video: tegra: add stride calculation for "complex" tiling requirements
some tegra engines place restrictions on the strides that are supported
for tiled surfaces. engines which do not have general dividers in their
addressing logic restrict strides to be an odd multiple of a power-of-two
number of tiles (where the odd multiple must be <= 15 and the power-of-two
must be <= 2**6).

extend tegra_dc_compute_stride to compute the smallest stride which meets
these requirements that is large enough to fit the requested surface
width and color depth.

Change-Id: Id66b629536f5e959a517b414f3d4bc8d9d80ddaf
Signed-off-by: Gary King <gking@nvidia.com>