tegra: hwmon: driver for nct1008 temperature monitor
this driver programs the temperature monitor to throttle the cpu
when the temperature crosses a certain level. the driver also
programs the sensor to power down the cpu if the temperature
still keeps on rising and crosses a certain threshold.

the thresholds should be set in the driver's platform data since
these may vary for each platform.

Change-Id: I76641c8981356ea8385a825f0b24c4f94f58b06a
Signed-off-by: Varun Wadekar <vwadekar@nvidia.com>